calendar_automator/requirements.txt
pytz
ipdb
python-dateutil 
termcolor
oauth2client
google-api-python-client